+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## unreleased
 * Add `isNetworkTokenized` to `AndroidPayCard` and `AndroidPayCardDetails`
 * Add support for PHP 7.4 (#262 thanks @slt)
+* Add `threeDSecurePassThru` params to PaymentMethod update
 
 ## 4.5.0
 * Add `PROCESSOR_DOES_NOT_SUPPORT_MOTO_FOR_CARD_TYPE` to validation errors